% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dst}
\alias{dst}
\title{Geodesic distance}
\usage{
dst(M)
}
\arguments{
\item{M}{Input adjacency matrix}
}
\value{
Matrix containing all the pairwise geodesic distances
}
\description{
Evaluate geodesic distance (shortest path) between all pairs of nodes in the network.
}
\examples{
dst(example_adjacency_matrix)
}
