% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_adjacency_matrix}
\alias{example_adjacency_matrix}
\title{Example Adjacency Matrix}
\format{A binary adjacency matrix representing links between nodes.}
\usage{
example_adjacency_matrix
}
\description{
Adjacency matrix of a 10 nodes random network for testing purposes
}
\keyword{datasets}
