% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lpY}
\alias{lpY}
\title{Network log-likelihood}
\usage{
lpY(Y, lpz, alpha, W)
}
\arguments{
\item{Y}{Adjacency matrix of the observed network}

\item{lpz}{Matrix containing the negative square root of the Euclidean distances between latent positions 
(output of \link[BLSM]{lpz_dist})}

\item{alpha}{Model variable \eqn{\alpha}}

\item{W}{BLSM Weights matrix of the observed network}
}
\value{
Log-likelihood of the observed network
}
\description{
Compute the log-likelihood of the whole observed network based on the
latent positions estimates and the model assumptions. See \link[BLSM]{BLSM} for more information.
}
