% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mlpY}
\alias{mlpY}
\title{Network (positive) log-likelihood}
\usage{
mlpY(avZ, Y, W)
}
\arguments{
\item{avZ}{Vector containing the \eqn{\alpha} value and the latent positions}

\item{Y}{Adjacency matrix of the observed network}

\item{W}{BLSM Weights matrix of the observed network}
}
\value{
Log-likelihood of the observed network
}
\description{
Compute the (positive) log-likelihood of the whole observed network based on the
latent positions estimates and the model assumptions. The inputs are slightly different from those of \link[BLSM]{lpY},
so the function basically applies some preprocessing before calling \link[BLSM]{lpY} and returning its value with the opposite sign.
}
