% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Z_up}
\alias{Z_up}
\title{Update step for the latent positions}
\usage{
Z_up(Y, Z, W, alpha, zdelta, mu_z, sd_z)
}
\arguments{
\item{Y}{Adjacency matrix of the observed network}

\item{Z}{Latent positions matrix}

\item{W}{BLSM Weights matrix of the observed network}

\item{alpha}{Model variable \eqn{\alpha}}

\item{zdelta}{Standard deviation of the Gaussian proposal for latent positions}

\item{mu_z}{Mean of the Gaussian prior distribution for latent positions}

\item{sd_z}{Standard deviation of the Gaussian prior distribution for latent positions}
}
\value{
Updated latent positions matrix
}
\description{
Accept/reject the proposals for the latent positions
}
