% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BLSM}
\alias{BLSM}
\alias{BLSM-package}
\title{Bayesian Latent Space Model}
\description{
BLSM
}
\details{
Add some info
}
\references{
AAAA
}
\author{
Alberto Donizetti <albe.donizetti@gmail.com>
}
