% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BLSM}
\alias{BLSM}
\alias{BLSM-package}
\title{Bayesian Latent Space Model}
\details{
R package allowing the computation of a Bayesian Latent Space Model for Complex Networks.
Given a network characterized by its adjacency and ... TBD
}
\references{
TBD
}
