% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsm.R
\name{estimate_latent_positions}
\alias{estimate_latent_positions}
\title{Core BLSM simulation}
\usage{
estimate_latent_positions(Y, W, procrustean = TRUE, k = 2, alpha = 2,
  nscan = 8 * 10^5, burn_in = 5 * 10^5, odens = 10^3, zdelta = 10,
  z_norm_prior_mu = 0, z_norm_prior_sd = 10, adelta = 0.3,
  a_exp_prior_a = 1, a_exp_prior_b = 1, dynamic_plot = FALSE,
  dynamic_circles = FALSE, ...)
}
\arguments{
\item{Y}{Adjacency matrix of the network}

\item{W}{Weight matrix of the network}

\item{procrustean}{Boolean to include/exclude (TRUE/FALSE) the Procrustean Transform step in the algorithm. Set TRUE by default.}

\item{k}{Space dimensionality}

\item{alpha}{Starting value of the \code{alpha} parameter}

\item{nscan}{Number of iterations}

\item{burn_in}{Burn-in value (starting iterations to be discared)}

\item{odens}{Thinning: only 1 iteration every \code{odens} will be sampled and stored in the output}

\item{zdelta}{Standard deviation of the Gaussian proposal for latent positions}

\item{z_norm_prior_mu}{Mean of the Gaussian prior distribution for latent positions}

\item{z_norm_prior_sd}{Standard deviation of the Gaussian prior distribution for latent positions}

\item{adelta}{The uniform proposal for \code{alpha} is defined on the \eqn{[-adelta,+adelta]} interval}

\item{a_exp_prior_a}{Shape parameter of the Gamma prior distribution for \code{alpha}. As the value is usually set to 1 the prior is an exponential distribution.}

\item{a_exp_prior_b}{Rate parameter of the Gamma prior distribution for \code{alpha}.}

\item{dynamic_plot}{Boolean to plot dinamically the simulated positions (one update every \code{odens} iterations)}

\item{dynamic_circles}{Boolean to add circles of radius \code{alpha} to the dynamic plots}

\item{{\dots}}{Additional parameters that can be passed to \link[BLSM]{plot_latent_positions}}
}
\value{
Returns a "BLSM object" (\code{blsm_obj}), i.e. a list containing:
\itemize{
\item \code{Alpha} {\code{alpha} values from the sampled iterations}
\item \code{Likelihood}{Log-likelihood values from the sampled iterations}
\item \code{Iterations}{Latent space coordinates from the sampled iterations. Latent positions are stored in a
3D array whose dimensions are given by (1: number of nodes, 2: space dimensionality, 3: number of iterations).
In the non-Procrustean framework the latent distances are given instead of the positions: another 3D array is returned, whose dimensions
are given by (1: number of nodes, 2: number of nodes, 3: number of iterations). The command needed in order to get the average values over the iterations for
either the positions or the distances is \code{rowMeans(blsm_obj$Iterations, dims=2)}.}
\item \code{StartingPositions}{Latent space coordinates right after the initialization step. In the non-Procrustean framework starting distances are given instead.}
\item \code{Matrix}{Original matrices of the network (adjacency and weights)}
\item \code{Parameters}{List of parameters specified during the call to \code{estimate_latent_positions}}
}
}
\description{
Run a simulation to obtain the positions of the network nodes in the latent space for each sampled iteration
}
