% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsm.R
\name{plot_traceplots_acf}
\alias{plot_traceplots_acf}
\title{BLSM traceplots and ACF}
\usage{
plot_traceplots_acf(blsm_obj, chosen_node = 1, coordinate = 1,
  chosen_pair = c(1, 2))
}
\arguments{
\item{blsm_obj}{Blsm object obtained through \link[BLSM]{estimate_latent_positions}}

\item{chosen_node}{Specified node for traceplot and autocorrelation function (Procrustean framework)}

\item{coordinate}{Specified coordinate dimension from the n-dimensional latent space}

\item{chosen_pair}{Specified pair of nodes for traceplot and autocorrelation function (non-Procrustean framework)}
}
\description{
Traceplots and autocorrelation functions for the \code{alpha} parameter and a selected node (or pair of nodes in the non-Procrustean framework)
}
