% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsm.R
\name{proc_crr}
\alias{proc_crr}
\title{Procrustean corresponding positions}
\usage{
proc_crr(Z, Z0)
}
\arguments{
\item{Z}{set of initial coordinates to be transformed}

\item{Z0}{set of reference coordinates}
}
\value{
Set of coordinates minimizing the distance between the initial configuration and the reference one
}
\description{
Given a set of starting coordinates, the function returns the Procrustean Transform of the initial points that minimizes 
the sum of squared positional difference from a set of reference coordinates. The (Euclidean) distances between a candidate 
configuration and the reference are evaluated by considering the couples of corresponding points.
}
